cr-backend/requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
cfgv==3.4.0
click==8.1.7
distlib==0.3.8
Django==4.2
django-jet-reboot==1.3.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
filelock==3.13.1
identify==2.5.35
mypy-extensions==1.0.0
nodeenv==1.8.0
packaging==23.2
pathspec==0.12.1
platformdirs==4.1.0
pre-commit==3.5.0
PyJWT==2.8.0
pytz==2023.3.post1
PyYAML==6.0.1
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.7.1
virtualenv==20.25.1
